move drag_drop_router from GUI to utils layer (#586)
drag_drop_router.py is pure Python with no Qt dependencies — it belongs
in the utils layer, not the GUI layer. Update all imports accordingly.
